+++ tgbotapi.core/src/commonTest/kotlin/dev/inmo/tgbotapi/types/MessageEntity/EntitiesTestText.kt
@@ -4,8 +4,8 @@ import dev.inmo.tgbotapi.types.MessageEntity.textsources.*
 import kotlin.test.assertTrue
 
 const val testText = "It is simple hello world with #tag and @mention"
-const val formattedV2Text = "It *_is_ ~__simple__~* hello world with \\#tag and @mention"
-const val formattedHtmlText = "It <b><i>is</i> <s><u>simple</u></s></b> hello world with #tag and @mention"
+const val formattedV2Text = "It *_is_ ~__simple__~* ||hello world|| with \\#tag and @mention"
+const val formattedHtmlText = "It <b><i>is</i> <s><u>simple</u></s></b> <span class=\"tg-spoiler\">hello world</span> with #tag and @mention"
 internal val testTextEntities = listOf(
     RawMessageEntity(
         "bold",
@@ -28,6 +28,11 @@ internal val testTextEntities = listOf(
         6
     ),
     RawMessageEntity(
+        "spoiler",
+        13,
+        11
+    ),
+    RawMessageEntity(
         "hashtag",
         30,
         4
@@ -43,9 +48,11 @@ fun TextSourcesList.testTextSources() {
     assertTrue (first() is RegularTextSource)
     assertTrue (get(1) is BoldTextSource)
     assertTrue (get(2) is RegularTextSource)
-    assertTrue (get(3) is HashTagTextSource)
+    assertTrue (get(3) is SpoilerTextSource)
     assertTrue (get(4) is RegularTextSource)
-    assertTrue (get(5) is MentionTextSource)
+    assertTrue (get(5) is HashTagTextSource)
+    assertTrue (get(6) is RegularTextSource)
+    assertTrue (get(7) is MentionTextSource)
 
     val boldSource = get(1) as BoldTextSource
     assertTrue (boldSource.subsources.first() is ItalicTextSource)